fix(tui/settings): file-size — rename dynamic_map_breadcrumb to dynamic_map_chrome
dynamic_map_draw.rs hit 412 LOC after the tab-chip contrast fix.
Rename dynamic_map_breadcrumb.rs → dynamic_map_chrome.rs (now hosts
both header chrome helpers: breadcrumb + tab_highlight_style) and
move tab_highlight_style + its 2 tests there. Imports updated.

dynamic_map_draw.rs back to 348 LOC, dynamic_map_chrome.rs at 164 LOC.

directory-tree.md updated for the rename.

Refs #908
