+++ src/tui/screens/settings/schema_tab/widgets/dynamic_map_chrome.rs
@@ -1,12 +1,37 @@
-//! Nested-editor breadcrumb helper for [`super::dynamic_map_draw`].
+//! Header-area chrome helpers for [`super::dynamic_map_draw`]:
+//! nested-editor breadcrumb + tab-chip highlight style.
 //!
 //! Split out of `dynamic_map_draw.rs` (#908) to keep both files under
-//! the 400-LOC guardrail. Pure string-shaping logic — easy to unit-test
+//! the 400-LOC guardrail. Pure presentation logic — easy to unit-test
 //! in isolation.
 
+use ratatui::style::{Modifier, Style};
+
+use crate::tui::theme::Theme;
 use crate::util::formatting::truncate_with_ellipsis;
 
 pub(super) const BREADCRUMB_SEP: &str = " → ";
+
+/// Tab-chip highlight style. Full orange chip when the SubtabStrip is
+/// the current chord target; muted underline-only chip otherwise.
+///
+/// Bright requires BOTH the outer caller treats this widget as the
+/// chord target (`focused=true`) AND its own focus is on the
+/// SubtabStrip. Caller folds those two bits into a single
+/// `chord_target` bool. Mirrors the dim-tab-strip pattern from the
+/// sidebar — only one chip on screen "shouts" at a time (#908).
+pub(super) fn tab_highlight_style(theme: &Theme, chord_target: bool) -> Style {
+    if chord_target {
+        Style::default()
+            .fg(theme.selection_fg)
+            .bg(theme.selection_bg)
+            .add_modifier(Modifier::BOLD)
+    } else {
+        Style::default()
+            .fg(theme.text_secondary)
+            .add_modifier(Modifier::UNDERLINED)
+    }
+}
 
 /// Build the nested-editor breadcrumb when `section_path` carries a
 /// parent's dotted path (3+ segments). Returns `None` for top-level
@@ -45,6 +70,49 @@ pub(super) fn nested_breadcrumb(
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[test]
+    fn tab_highlight_full_chip_when_chord_target() {
+        // #908 contrast — bright when caller passes chord_target=true.
+        let theme = Theme::dark();
+        let style = tab_highlight_style(&theme, true);
+        assert_eq!(
+            style.bg,
+            Some(theme.selection_bg),
+            "must paint selection bg"
+        );
+        assert_eq!(
+            style.fg,
+            Some(theme.selection_fg),
+            "must paint selection fg"
+        );
+        assert!(
+            style.add_modifier.contains(Modifier::BOLD),
+            "must be BOLD when chord target"
+        );
+    }
+
+    #[test]
+    fn tab_highlight_dim_underline_when_not_chord_target() {
+        // #908 contrast — dim once chord target is elsewhere (focus
+        // descended into an EntryField, or widget itself is unfocused).
+        let theme = Theme::dark();
+        let style = tab_highlight_style(&theme, false);
+        assert_eq!(style.bg, None, "must NOT paint bg when not chord target");
+        assert_eq!(
+            style.fg,
+            Some(theme.text_secondary),
+            "must use text_secondary fg when not chord target"
+        );
+        assert!(
+            !style.add_modifier.contains(Modifier::BOLD),
+            "must NOT be BOLD when not chord target"
+        );
+        assert!(
+            style.add_modifier.contains(Modifier::UNDERLINED),
+            "must be UNDERLINED to keep the active-tab signal"
+        );
+    }
 
     #[test]
     fn returns_none_for_top_level_section() {
