+++ src/tui/screens/settings/schema_tab/widgets/mod.rs
@@ -8,7 +8,7 @@
 
 pub mod clock;
 pub mod dynamic_map;
-mod dynamic_map_breadcrumb;
+mod dynamic_map_chrome;
 mod dynamic_map_draw;
 pub mod dynamic_rows;
 mod dynamic_rows_draw;
